requirements.txt
backcall==0.1.0
cycler==0.10.0
decorator==4.3.2
future==0.17.1
gps==3.9
imageio==2.5.0
imutils==0.5.2
ipdb==0.11
ipython==7.3.0
ipython-genutils==0.2.0
iso8601==0.1.12
jedi==0.13.3
kiwisolver==1.1.0
matplotlib==3.0.3
networkx==2.3
numpy==1.16.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.3.1
pynmea2==1.15.0
pyparsing==2.4.0
pyserial==3.4
python-dateutil==2.8.0
PyWavelets==1.0.3
PyYAML==5.1.1
scikit-image==0.15.0
scipy==1.3.0
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
